Fix GCC5 incompatibility
C++11 ABI is not used by everything yet.

https://developerblog.redhat.com/2015/02/05/gcc5-and-the-c11-abi/

Correct GCC5 fix

When the original -D_GLIBCXX_USE_CXX11_ABI=0 fix was suggested, it
made sense to me. I thought the fuse errors (including "Invalid
characters encountered") were improper string exchange between
CipherShed and libfuse, due to an ABI difference. Upon further
investigation, I have discovered that the fuse issue was due to
failed command execution.

Platform/Unix/Process.cpp uses TrueCrypt's own foreach implementation
to simplify loops. A foreach loop is used to copy the data from
std::string instances to raw C-strings to pass to another function. A
side effect of the foreach loop is that new strings are constructed
from the original strings, so when CipherShed stores the C-string
pointers, it's storing the pointers of the copies instead of the
originals. This is error-prone because these copies are immediately
deconstructed, leaving us with invalid pointers. It's only
coincidence that this side effect does not cause issues when built
with the old string implementation (which is what
-D_GLIBCXX_USE_CXX11_ABI=0 was making us use). Simply using a real
for-loop fixes the re-construction issue. This also compiles with
both GCCs and both ABIs.

However, other usage of foreach needs to be inspected for similar
bugs. Foreach may need to be discarded entirely, since a proper
implementation may require us to rewrite each usage of it. At that
point, we might as well just use for loops.

Also, after some thinking, it should not be our responsibility to set
_GLIBCXX_USE_CXX11_ABI, since it may only be necessary on certain
platforms. Platform configuration will decide the value of this.