wx3.0: Fix strange volume size validator behavior
VolumeSizeWizardPage::IsValid() is called from a text-change event
(EVT_TEXT) handler, and uses IsEmpty() on a wxTextCtrl as part of
determining if the input is valid. It appears that in wxWidgets 3.0,
the data that IsEmpty relies on is not updated until *after*
text-change event handlers are processed, so the character count is
always behind by 1 character, making the wizard think valid sizes are
invalid. Using GetValue().empty() still works, but these functions
will eventually need to be re-implemented to process the data given
to the event handler, instead.